refactor: 도메인별로 타입 분리
