+++ frontend/src/types/space-content.type.ts
@@ -1,0 +1,10 @@
+export interface SpaceContent {
+  id: number;
+  contentType: string;
+  spaceId: number;
+}
+
+export interface CreateSpaceContentInput {
+  contentType: string;
+  spaceId: number;
+}
